@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^13.3.11",
-    "@angular/animations": "^13.4.0",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/router": "^13.4.0",
+    "@angular-devkit/schematics": "^14.2.12",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@fortawesome/fontawesome-free": "^5.12.1",
     "@types/highlight.js": "^10.1.0",
     "bulma": "^0.9.3",
@@ -31,10 +31,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.11",
-    "@angular/cli": "^13.3.11",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/language-service": "^13.4.0",
+    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^13.9.1",
