@@ -7,7 +7,7 @@
 import { Component, OnInit, Injectable } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
-import { FormGroup,  FormBuilder, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup,  UntypedFormBuilder, UntypedFormControl, Validators } from '@angular/forms';
 
 import { ClipboardService } from './clipboard.service';
 import { componentIdValidator, urlValidator, noPathOrSpaceValidator } from './formvalidators';
@@ -23,7 +23,7 @@ import { makeDesktopEntryData } from './makeauxdata';
 
 @Injectable()
 export class AddonComponent implements OnInit {
-    cptForm: FormGroup;
+    cptForm: UntypedFormGroup;
     finalCptId: string;
 
     metadataLicenses: any;
@@ -36,7 +36,7 @@ export class AddonComponent implements OnInit {
     dataMetainfo: string;
     dataMesonValidate: string;
 
-    constructor(private fb: FormBuilder,
+    constructor(private fb: UntypedFormBuilder,
                 private http: HttpClient,
                 public clipboard: ClipboardService) {
         this.dataGenerated = false;
@@ -61,8 +61,8 @@ export class AddonComponent implements OnInit {
 
             metadataLicense: ['', Validators.required ],
             rbLicenseMode: [''],
-            simpleProjectLicense: new FormControl({value: '', disabled: false}),
-            complexProjectLicense: new FormControl({value: '', disabled: true}),
+            simpleProjectLicense: new UntypedFormControl({value: '', disabled: false}),
+            complexProjectLicense: new UntypedFormControl({value: '', disabled: true}),
 
             cptIcon: ['', noPathOrSpaceValidator() ],
 
