@@ -7,7 +7,7 @@
 import { Component, OnInit, Injectable } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
-import { FormGroup,  FormBuilder, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup,  UntypedFormBuilder, UntypedFormControl, Validators } from '@angular/forms';
 
 import { ClipboardService } from './clipboard.service';
 import { componentIdValidator, urlValidator, desktopEntryValidator,
@@ -25,7 +25,7 @@ import { makeDesktopEntryData } from './makeauxdata';
 
 @Injectable()
 export class GUIAppComponent implements OnInit {
-    cptForm: FormGroup;
+    cptForm: UntypedFormGroup;
     finalCptId: string;
 
     metadataLicenses: any;
@@ -48,7 +48,7 @@ export class GUIAppComponent implements OnInit {
     dataMesonL10N: string;
     dataMesonMItoDE: string;
 
-    constructor(private fb: FormBuilder,
+    constructor(private fb: UntypedFormBuilder,
                 private http: HttpClient,
                 public clipboard: ClipboardService) {
         this.dataGenerated = false;
@@ -81,8 +81,8 @@ export class GUIAppComponent implements OnInit {
             cptId: ['', [ Validators.required, Validators.minLength(4), componentIdValidator() ]],
             metadataLicense: ['', Validators.required ],
             rbLicenseMode: [''],
-            simpleProjectLicense: new FormControl({value: '', disabled: false}),
-            complexProjectLicense: new FormControl({value: '', disabled: true}),
+            simpleProjectLicense: new UntypedFormControl({value: '', disabled: false}),
+            complexProjectLicense: new UntypedFormControl({value: '', disabled: true}),
 
             primaryScreenshot: ['', urlValidator() ],
             extraScreenshot1: ['', urlValidator() ],
